example/simple-project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dapr==1.13.1
requests==2.31.0
pydantic==2.5.0
